--- src/block-management/block-holder.scss
@@ -1,34 +1,0 @@
-/** @format */
-
-@import 'block-styles.scss';
-
-.blockHolder {
-	flex: 1 1 auto;
-}
-
-.blockContainer {
-	background-color: white;
-	padding-left: 16;
-	padding-right: 16;
-	padding-top: 12;
-	padding-bottom: 12;
-}
-
-.blockContainerFocused {
-	background-color: white;
-	padding-left: 16;
-	padding-right: 16;
-	padding-top: 12;
-	padding-bottom: 0; // will be flushed into inline toolbar height
-}
-
-.blockTitle {
-	background-color: grey;
-	padding-left: 8;
-	padding-top: 4;
-	padding-bottom: 4;
-}
-
-.aztec_container {
-	flex: 1;
-}
