--- src/block-management/block-styles.scss
@@ -1,9 +1,0 @@
-/** @format */
-
-.blockCode {
-	font-family: $default-monospace-font;
-}
-
-.blockText {
-	min-height: 50;
-}
