--- src/block-management/inline-toolbar/actions.js
@@ -1,12 +1,0 @@
-/**
- * @format
- * @flow
- */
-
-const InlineToolbarActions = {
-	UP: 1,
-	DOWN: 2,
-	DELETE: 3,
-};
-
-export default InlineToolbarActions;
