--- src/components/readable-content-view.scss
@@ -1,8 +1,0 @@
-.container {
-	align-items: center;
-}
-
-.centeredContent {
-	width: 100%; 
-	max-width: 580;
-}
